pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llmring-py"
version = "0.2.2"
description = "Unified Python interface for OpenAI, Anthropic, Google, and Ollama LLMs with optional model registry and usage tracking"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Juan Reyero", email = "juan@juanreyero.com" }
]
maintainers = [
    { name = "Juan Reyero", email = "juan@juanreyero.com" }
]
keywords = ["llm", "ai", "openai", "anthropic", "claude", "gpt", "gemini", "ollama", "api"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.10"

# Runtime dependencies
dependencies = [
    "anthropic>=0.18.0",
    "openai>=1.0.0",
    "google-genai>=0.1.0",
    "httpx>=0.24.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.8.0",
    "ollama>=0.4.0",
    "psycopg2-binary>=2.9.0",
    "beautifulsoup4>=4.12.0",
    "aiosqlite>=0.19.0",
    "pgdbm>=0.1.0",
]

[project.urls]
Homepage = "https://juanreyero.com/open/llmring/"
Documentation = "https://github.com/juanre/llmring"
Repository = "https://github.com/juanre/llmring"
Issues = "https://github.com/juanre/llmring/issues"

# (removed duplicate [project] header)

[project.scripts]
llmring = "llmring.cli:main"

[project.optional-dependencies]
postgres = [
    "pgdbm>=0.1.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-mock>=3.10",
    "black>=23.0",
    "isort>=5.0",
    "pyright>=1.1",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pillow>=10.0.0",  # For image examples
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "claude-code-sdk>=0.0.10",
    "pdf2image>=1.17.0",
    "pillow>=11.2.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "reportlab>=4.4.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/llmring"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.hatch.build]
include = [
    "src/llmring/**/*.py",
    "src/llmring/**/*.sql",
    "src/llmring/migrations/*.sql",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "llm: marks tests as LLM-related",
    "unit: marks tests as unit tests",
    "integration: marks tests as integration tests",
    "slow: marks tests as slow",
    "db: marks tests as requiring database",
    "google: marks tests as Google-specific",
    "ollama: marks tests as Ollama-specific",
]

[tool.uv.sources]
pgdbm = { path = "../pgdbm" }

